% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_yhat_mat.R
\name{get_yhat_mat}
\alias{get_yhat_mat}
\title{Sample from the VBART... whatever that means}
\usage{
get_yhat_mat(vbart_obj)
}
\arguments{
\item{vbart_obj}{}
}
\value{
A matrix of results.
}
\description{
This function runs the VBART sampler.
}
\examples{
vbart_sampler(y, 1000, 50, 2, lon, lat)
}
