% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_vbart.R
\name{predict_vbart}
\alias{predict_vbart}
\title{Predict from the VBART... whatever that means}
\usage{
predict_vbart(vbart_obj, new_lon, new_lat)
}
\arguments{
\item{y}{A vector of data points.}

\item{n_samples}{A scalar giving the number of samples.}

\item{n_trees}{A scalar giving the number of trees.}

\item{delta}{A scalar parameter used to parameterize the underlying Poisson distribution.}

\item{lon}{A vector of longitudes.}

\item{lat}{A vector of latitudes.}
}
\value{
A list of results.
}
\description{
This function runs the VBART prediction.
}
\examples{
vbart_sampler(y, 1000, 50, 2, lon, lat)
}
