% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vbart_sampler.R
\name{vbart_sampler}
\alias{vbart_sampler}
\title{Sample from the VBART... whatever that means}
\usage{
vbart_sampler(y, n_samples, n_trees, delta, lon, lat)
}
\arguments{
\item{y}{A vector of data points.}

\item{n_samples}{A scalar giving the number of samples.}

\item{n_trees}{A scalar giving the number of trees.}

\item{delta}{A scalar parameter used to parameterize the underlying Poisson distribution.}

\item{lon}{A vector of longitudes.}

\item{lat}{A vector of latitudes.}
}
\value{
A list of results.
}
\description{
This function runs the VBART sampler.
}
\examples{
vbart_sampler(y, 1000, 50, 2, lon, lat)
}
